requirements.txt
faster-whisper>=1.1.0
PyAudio==0.2.14
keyboard==0.13.5
numpy>=1.20.0
Pillow>=10.0.0
pyperclip==1.8.2
pystray==0.19.5
